pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-network_forensics"
version = "0.1.0"
description = "Network Forensics environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.2",
    # Environment-specific dependencies
    "scapy>=2.5.0",
    "networkx>=3.0",
    "pydantic>=2.0",
    "numpy>=1.24.0",
    "faker>=20.0.0",
    "pytest>=8.0.0",
    "gradio>=4.0.0",
    "huggingface-hub>=0.19.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m network_forensics.server.app
server = "network_forensics.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["network_forensics", "network_forensics.server"]
package-dir = { "network_forensics" = ".", "network_forensics.server" = "server" }